Make the syncer into a child app, rename accordingly
This repositions the app as a child app for a new parent so that one can have many instances of this pairing of bulbs to LEDs.  The file has been renamed to show that it is a child app now.  The definition has been updated with a link to the relevant thread and source repository for semi-automated updates.  New logic for naming the child apps and overriding the default name.